+++ CHANGES.rst
@@ -66,6 +66,11 @@ gaia
 gama
 ^^^^
 
+gemini
+^^^^^^
+
+- initial support for the Gemini archive [#1596]
+
 heasarc
 ^^^^^^^